+++ netket/operator/__init__.py
@@ -19,7 +19,9 @@ from ._local_operator import LocalOperator
 from ._graph_operator import GraphOperator
 from ._pauli_strings import PauliStrings
 from ._lazy import Adjoint, Transpose, Squared
-from ._hamiltonian import Ising, Heisenberg, BoseHubbard
+from ._heisenberg import Heisenberg
+from ._ising import Ising
+from ._bose_hubbard import BoseHubbard
 
 from ._abstract_super_operator import AbstractSuperOperator
 from ._local_liouvillian import LocalLiouvillian
